Hide a section until product is generally available
We created a set of sample apps/components, but unfortunately they cannot be loaded in production until the new bits release.
Rather than delete the section or hold it in a private branch, we would like to just comment the section out and will uncomment and expose with that release in a few weeks.