Notification on bot default access configuration
Adding a "Warning" block on the new bot creation documentation per discussion. We need to inform users of the default bot access configurations and point to their documentation pages. In addition, we will include a message in the product itself, and in the future, add references to the DLP documentation as well.