Update setup.md
Updated the Installing Updates section to include running the admin sync flows with Full Inventory on.  This would have really saved me some time if I knew this before I upgraded to the June release.  The Flow Connections column on the Flow table are empty after Core solution upgrade which means that the Impact Analysis of the DLP editor does not work - not impacted flow or apps are returned when there should be.