Update about-powerapps-perapp.md
Re-introducing the previous mitigation step for Are users with a disabled account in an environment able to use per app plan? BUT the reintroduction includes a couple of edits that make the steps viable. E.g. disabled users must be added to a security group and then the app must be shared with the security group. One cannot share with disabled users directly.