Revised the HEAT article
Completely redid the HEAT article based on the blog post. I need to figure out how to make the images into hyperlinks that go to the YouTube videos. Pranav needs to check every link and give me updates for the broken ones. I left the old doc here for now (as heat-old-delete-me) but that needs to be deleted from the set before we publish.