Update overview-alm.md
Revising for style. Also removing "to enable some of these features and tools" because it seems confusing to me to mention this here without providing specifics as many of the tools and services *don't* strictly require Dataverse and have other purposes, such as a source control system (e.g. Git/GitHub) or Azure DevOps.