Fix Broken Links
Links that reference /power-platform/admin/<target>.md were still broken. The absolute paths seem to be correct, but the links are keeping the .md and causing 404 error. Since the md files are in same repo, the absolute paths should work so let me know if I need to adjust this PR.