Update function-concatenate.md
I'm going through all of the Power FX doco to learn it and create a reference app. I noticed that the examples for Concat were doing an & ", " instead of using the separator parameter. I have updated them to use the separator parameter after testing that it works. This is my first pull request, so let me know if I have done it wrong. Thanks