Update function-dateadd-datediff.md
I think the examples must have been written before 2014. The dates in the DateDiff examples give negative numbers if you enter them as is. I swapped them around so Now() is the second parameter and changed the date to 2022. Now it should always give the correct number.

I know, it is picky... :)