Update managed-environment-overview.md
The landing page seems to contradict the Known Limitations page - as the feature is new, perhaps it's worth advising of limitations up front so customers know whether it is an option worth exploring. I've expanded on the contradiction and added a link to the Known Limitations page. Break space between the two links.