Remove incorrect deduction of an orphan app
Line 93 incorrectly states that an app is orphaned if it doesn't have a "Last opened" date value. This is not the case. An app may have an owner and not be orphaned, even if date value for last opened isn't present. I don't believe there's any report in PPAC that reliably displays orphaned canvas apps as the original creator owner is always displayed in tenant level analytics as well as when looking at Power Apps behind an environment in PPAC, even if the creator owner is no longer in AAD. The only reliable ways to distinguish orphaned apps are with the CoE Starter Kit's Set App Permissions and Power Platfrom Admin View apps.

The concept of an orphaned app also doesn't apply to model-driven apps as access to them is defined by security roles. In that sense, this Learn page is also misleading when it talks about orpahned apps.