Update manage-principalobjectaccess-storage.md
Fixed the topic of the cascade relationship behavior from direct sharing to indirect sharing, since this is the way how POA works for relationship behaviors.

Described that manual deletion in POA table is not supported and the proper way is to adjust security model and revoke access when needed. This will avoid wrong expectations when creating a support ticket.