Update manage-date-and-time.md
The current documentation is wrong, in that it states datetime values are stored as UTC and you must add the user's current time zone offset to return a value in the user's current time zone. However, datetime values, including Now(), are stored / output in the current time zone. These changes to the article make the documentation match how Copilot Studio works.