Update connector-endpoint-filtering.md
Users of Power Automate/ Power Apps have been confused because they could not find documentation that confirms that these rules also apply to SQL server connectors accessing Azure Synapse datawarehouses.  Customers have been confused and not aware that Azure Synapse is a SQL database that happens to be a datawarehouse.