Update powerapps-flow-licensing-faq.md for Extended Use Rights
I have added an new FAQ related to Extended Use Rights of Power Apps with Microsoft 365. This is an important FAQ that will help readers understand ways to assess the users and apps under extended use rights and is required starting first week of April 2025 as the enforcement begins. Kindly review and approve