Update connect-exchange-online.md
Update settings access steps for new and classic PPAC. 
Refresh updated image in PPAC.
Remove the doc section "## Test email configuration for all mailboxes associated with an email server profile", as there is no command for "Test & Enable Mailboxes" when select a sever profile. This section of the documentation is outdated.