Updated images to not show legacy PPAC
I also made a few updates. I updated the title because these admins are not going to "disable," they are going to review connectors already disabled by default and determine if they want to enable any of them. Since it's a how-to article, I added prerequisites. a little redundant, but it matches the article-type template more closely and reminds readers that this feature is for a very limited audience. The images are zoomed in a bit so as not to show legacy PPAC as much. The quality is not the best—so I recommend adding new ones when possible. Sometimes we just remove images, but in this case I want the admins to see what shows up for their tenants—it's unique.