Refine Global Secure Access setup steps
I walked through the configuration steps in my test environment now that I can access the feature in PPAC and updated the environment and environment group steps so they align with one another better. For environment groups we could even just start the first step on the Configuation pane because that's where there's the small departure from the environment steps. Another option would be to create tabs. The reader could select "Environment" or "Environment group" to see the related instructions. Please double-check before we publish to ensure they read well to you and are accurate.